Refactor code to reduce duplicate code
Implement a way so that areguments and options can be used by both directives and resolvers.

 - Create Option class to be used by both resolvers and directives
 - Create OptionArgument node for both resolvers and directives to extend
 - Refactor code for ReferenceNode and DirectiveNode to use OptionArgument class
 - Remove DirectiveOption and NodeOption
 - Create OptionArgument for validating and fetching arguments and options
 - Refactor ReferenceResolver and DirectiveProcessor to use OptionArgumentProcessor
 - Update the reference resolve method to use the node and not an explicit list of arguments
