+++ core/src/test/java/gyro/core/scope/ValueReferenceResolver.java
@@ -16,8 +16,6 @@
 
 package gyro.core.scope;
 
-import java.util.List;
-
 import gyro.core.Type;
 import gyro.core.reference.ReferenceResolver;
 import gyro.lang.ast.value.ReferenceNode;
@@ -26,8 +24,8 @@ import gyro.lang.ast.value.ReferenceNode;
 public class ValueReferenceResolver extends ReferenceResolver {
 
     @Override
-    public Object resolve(ReferenceNode node, Scope scope, List<Object> arguments) {
-        return arguments.isEmpty() ? null : arguments.get(0);
+    public Object resolve(ReferenceNode node, Scope scope) {
+        return getArgument(scope, node, Object.class, 1);
     }
 
 }
