+++ lang/src/main/java/gyro/lang/ast/value/ReferenceNode.java
@@ -17,61 +17,38 @@
 package gyro.lang.ast.value;
 
 import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
 import gyro.lang.ast.Node;
 import gyro.lang.ast.NodeVisitor;
+import gyro.lang.ast.OptionArgumentNode;
 import gyro.lang.filter.Filter;
 import gyro.parser.antlr4.GyroParser;
 import gyro.util.ImmutableCollectors;
 
-public class ReferenceNode extends Node {
+public class ReferenceNode extends OptionArgumentNode {
 
-    private final List<Node> arguments;
     private final List<Filter> filters;
-    private final List<ReferenceOption> options;
 
-    public ReferenceNode(List<Node> arguments, Collection<Filter> filters, List<ReferenceOption> options) {
-        super(null);
+    public ReferenceNode(List<Node> arguments, Collection<Filter> filters, List<Option> options) {
+        super(arguments, options);
 
-        this.arguments = ImmutableList.copyOf(Preconditions.checkNotNull(arguments));
         this.filters = ImmutableList.copyOf(Preconditions.checkNotNull(filters));
-        this.options = ImmutableList.copyOf(Preconditions.checkNotNull(options));
     }
 
     public ReferenceNode(GyroParser.ReferenceContext context) {
         super(Preconditions.checkNotNull(context));
 
-        this.arguments = Optional.ofNullable(context.IDENTIFIER())
-            .map(Node::create)
-            .map(Collections::singletonList)
-            .orElseGet(() -> Node.create(context.value()));
-
         this.filters = context.filter()
             .stream()
             .map(Filter::create)
             .collect(ImmutableCollectors.toList());
-
-        this.options = context.option()
-            .stream()
-            .map(ReferenceOption::new)
-            .collect(ImmutableCollectors.toList());
-    }
-
-    public List<Node> getArguments() {
-        return arguments;
     }
 
     public List<Filter> getFilters() {
         return filters;
-    }
-
-    public List<ReferenceOption> getOptions() {
-        return options;
     }
 
     @Override
